+++ src/redux/actions/categoryActions.js
@@ -1,5 +1,7 @@
 import axios from "axios";
+import Swal from "sweetalert2";
 import { categoryService } from "../../api/services/categoryService";
+import { subCategoryService } from "../../api/services/subCategoryService";
 import * as actionTypes from "./actionTypes";
 
 export const changeCategory = (category) => (dispatch) => {
@@ -14,8 +16,12 @@ function getCategoriesSuccess(categories) {
 }
 
 export const getCategories = () => (dispatch) => {
-  let url = "https://localhost:5001/member/api/Categories?page=1";
-  return fetch(url)
+  let url = "https://localhost:5001/admin/api/Categories?page=1";
+  return fetch(url, {
+    headers: {
+      Authorization: `Bearer ${localStorage.getItem("token")}`,
+    },
+  })
     .then((response) => response.json())
     .then((result) => dispatch(getCategoriesSuccess(result)));
 };
@@ -25,15 +31,53 @@ export function getSubCategoriesSuccess(subCategories) {
 }
 
 export const getSubCategories = () => (dispatch) => {
-  let url = "https://localhost:5001/member/api/subCategories?page=1";
-  return fetch(url)
+  let url = "https://localhost:5001/admin/api/subCategories?page=1";
+  return fetch(url, {
+    headers: {
+      Authorization: `Bearer ${localStorage.getItem("token")}`,
+    },
+  })
     .then((response) => response.json())
-    .then((result) => dispatch(getSubCategoriesSuccess(result)));
+    .then((result) => {
+      dispatch(getSubCategoriesSuccess(result));
+    });
 };
 
 export const deleteCategory = (id) => (dispatch) => {
-  categoryService.deleteCategory(id).then(() => {
-    getCategories()(dispatch);
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire("Deleted!", "Your file has been deleted.", "success");
+      categoryService.deleteCategory(id).then(() => {
+        getCategories()(dispatch);
+      });
+    }
+  });
+};
+
+export const deleteSubCategory = (id) => (dispatch) => {
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire("Deleted!", "Your file has been deleted.", "success");
+      subCategoryService.deleteSubCategory(id).then(() => {
+        getSubCategories()(dispatch);
+      });
+    }
   });
 };
 
@@ -45,10 +89,6 @@ export const postCategory = (category) => (dispatch) => {
   let url = "https://localhost:5001/admin/api/Categories";
   var formData = new FormData();
   for (let property in category) {
-    if (property === "image") {
-      formData.append("photo", category[property]);
-      break;
-    }
     formData.append(property, category[property]);
   }
 
@@ -56,12 +96,15 @@ export const postCategory = (category) => (dispatch) => {
     .post(`${url}`, formData, {
       headers: {
         "Content-Type": "multipart/form-data",
+        Authorization: `Bearer ${localStorage.getItem("token")}`,
       },
     })
-    .then((response) => {
-      console.log(response);
-      dispatch(postCategorySuccess);
-      dispatch(getCategories);
+    .then(({ data }) => {
+      dispatch({
+        type: actionTypes.POST_SUB_CATEGORY_SUCCESS,
+        payload: data,
+      });
+      getCategories()(dispatch);
     })
     .catch((error) => console.log(error));
 };
@@ -82,32 +125,40 @@ export const editCategory = (id, category) => (dispatch) => {
   });
 };
 
-function postSubCategoriesSuccess(categories) {
-  return { type: actionTypes.POST_SUB_CATEGORY_SUCCESS, payload: categories };
-}
+export const editSubCategory = (id, subCategory) => (dispatch) => {
+  const formData = new FormData();
 
-export const postSubCategory = (subCategory) => (dispatch) => {
-  let url = "https://localhost:5001/admin/api/subcategories";
-  var formData = new FormData();
+  formData.append("photo", subCategory.photo);
   for (let property in subCategory) {
-    if (property === "image") {
-      console.log(property);
-      formData.append("photo", subCategory[property]);
-      break;
-    }
     formData.append(property, subCategory[property]);
   }
 
+  subCategoryService.putSubCategory(formData, id).then(() => {
+    getSubCategories()(dispatch);
+  });
+};
+
+export const postSubCategory = (subCategory, push) => (dispatch) => {
+  let url = "https://localhost:5001/admin/api/subcategories";
+  var formData = new FormData();
+  for (let property in subCategory) {
+    formData.append(property, subCategory[property]);
+  }
   axios
     .post(`${url}`, formData, {
       headers: {
         "Content-Type": "multipart/form-data",
+
+        Authorization: `Bearer ${localStorage.getItem("token")}`,
       },
     })
-    .then((response) => {
-      console.log(response);
-      dispatch(postSubCategoriesSuccess);
-      dispatch(getSubCategories);
+    .then(({ data }) => {
+      dispatch({
+        type: actionTypes.POST_SUB_CATEGORY_SUCCESS,
+        payload: data,
+      });
+      getSubCategories()(dispatch);
+      push("/subcategories");
     })
     .catch((error) => console.log(error));
 };
